Small change for undefined variables in timelike shower.
Also note another small update yesterday, which was not sent on, 
since it involved the Python interface, which apparently is too big 
to be mailed as a change. At the same time another few very minor 
changes. 

It seems the program is working now; at least it survives several
sanity checks. But do keep on prodding. If all is well we should 
release in a few days. 



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/pythia8/pythia82/trunk@487 ffdf98f5-c1f1-482d-8be8-39a3f79b3492